# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Products"]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      }
    },
    "/api/products/count/category/{categoryId}": {
      "get": {
        "operationId": "ProductsController_countByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      }
    },
    "/api/products/top-sales": {
      "get": {
        "operationId": "ProductsController_getTopProducts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Products"]
      }
    },
...
    "/api/transportings/order/{orderId}": {
      "get": {
        "operationId": "TransportingsController_findByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "example": "507f1f77bcf86cd799439012",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transporting record found"
          },
          "404": {
            "description": "Transporting record not found"
          }
        },
        "summary": "Get transporting record by order ID",
        "tags": ["transportings"]
      }
    },
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      },
      "get": {
        "operationId": "OrdersController_find",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
    "/api/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
    "/api/orders/{id}/mark-paid": {
      "patch": {
        "operationId": "OrdersController_markAsPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
    "/api/orders/{id}/mark-shipped": {
      "patch": {
        "operationId": "OrdersController_markAsShipped",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
...
    "/api/orders/{id}/refund": {
      "patch": {
        "operationId": "OrdersController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Orders"]
      }
    },
...
    "/api/payment/initiate": {
      "post": {
        "operationId": "PaymentController_initiatePayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateZarinpalPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Payment"]
      }
    },
    "/api/payment/verify": {
      "post": {
        "operationId": "PaymentController_verifyPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyZarinpalPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Payment"]
      }
    },
    "/api/payment/process-refund": {
      "post": {
        "operationId": "PaymentController_processRefund",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRefundZarinpalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Payment"]
      }
    },
...
    "/api/payment/callback": {
      "get": {
        "operationId": "PaymentController_handleCallback",
        "parameters": [
          {
            "name": "Authority",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Payment"]
      }
    },
...
  },
  "info": {
    "title": "practice",
    "description": "API documentation for my practice app",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "StockDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "Quantity of product in stock",
            "example": 500,
            "minimum": 0
          }
        },
        "required": ["quantity"]
      },
      "VariantOptionDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Value of the variant option",
            "example": "50 kg"
          },
          "priceModifier": {
            "type": "number",
            "description": "Price modifier for this option (in IRR)",
            "example": 0
          }
        },
        "required": ["value"]
      },
      "VariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the variant",
            "example": "Packaging"
          },
          "options": {
            "description": "List of variant options",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantOptionDto"
            }
          }
        },
        "required": ["name", "options"]
      },
      "ImageDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "URL of the product image",
            "example": "https://example.com/cement-bag.jpg"
          }
        },
        "required": ["url"]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the product",
            "example": "Cement Type 2"
          },
          "slug": {
            "type": "string",
            "description": "SEO-friendly unique slug for the product",
            "example": "cement-type-2"
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit (SKU) code",
            "example": "CEM-T2-001"
          },
          "basePrice": {
            "type": "number",
            "description": "Base price of the product (in IRR)",
            "example": 2000000,
            "minimum": 0
          },
          "companyId": {
            "type": "string",
            "description": "MongoDB ObjectId of the supplier company",
            "example": "507f1f77bcf86cd799439011"
          },
          "categories": {
            "description": "List of category IDs",
            "example": ["507f1f77bcf86cd799439012"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "example": "High-strength Type 2 cement for concrete structures",
            "maxLength": 500
          },
          "stock": {
            "description": "Stock information for the product",
            "allOf": [
              {
                "$ref": "#/components/schemas/StockDto"
              }
            ]
          },
          "variants": {
            "description": "Product variants",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          },
          "attributes": {
            "type": "object",
            "description": "Custom attributes of the product",
            "example": {
              "strength": "42.5 MPa",
              "settingTime": "3 hours"
            }
          },
          "tags": {
            "description": "Tags for filtering and search",
            "example": ["cement", "construction"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "images": {
            "description": "Images associated with the product",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageDto"
            }
          },
          "comments": {
            "description": "User comments about the product",
            "example": ["Great quality!", "Fast delivery"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "number",
            "description": "Average rating of the product (1 to 5)",
            "example": 4.5,
            "minimum": 1,
            "maximum": 5
          },
          "status": {
            "type": "string",
            "description": "Status of the product",
            "enum": ["active", "inactive", "draft", "archived"],
            "example": "draft"
          }
        },
        "required": ["name", "slug", "sku", "basePrice", "companyId", "stock"]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "MongoDB ObjectId of the product",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": ["id"]
      },
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "MongoDB ObjectId of the product",
            "example": "507f1f77bcf86cd799439011"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the product",
            "example": 2,
            "minimum": 1
          }
        },
        "required": ["productId", "quantity"]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "MongoDB ObjectId of the user",
            "example": "507f1f77bcf86cd799439011"
          },
          "items": {
            "description": "List of order items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "totalPrice": {
            "type": "number",
            "description": "Total price of the order (in IRR)",
            "example": 1500000,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "Status of the order",
            "enum": [
              "pending",
              "paid",
              "shipped",
              "delivered",
              "completed",
              "refunded",
              "canceled",
              "failed"
            ],
            "example": "pending"
          },
          "shippingAddress": {
            "type": "string",
            "description": "Shipping address for the order",
            "example": "123 Street, Tehran, Iran"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method used for the order",
            "example": "Credit Card"
          },
          "companyId": {
            "type": "string",
            "description": "MongoDB ObjectId of the supplier company",
            "example": "507f1f77bcf86cd799439011"
          },
          "transportId": {
            "type": "string",
            "description": "MongoDB ObjectId of the transport record",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": ["userId", "items", "totalPrice", "status", "companyId"]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "MongoDB ObjectId of the order",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": ["id"]
      },
...
      "InitiateZarinpalPaymentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Payment amount in IRR",
            "example": 100000
          },
          "callbackUrl": {
            "type": "string",
            "description": "Callback URL for payment",
            "example": "http://localhost:3000/payment/callback"
          },
          "description": {
            "type": "string",
            "description": "Payment description",
            "example": "Payment for order #123"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile number of the payer",
            "example": "09123456789"
          },
          "email": {
            "type": "string",
            "description": "Email address of the payer",
            "example": "user@example.com"
          }
        },
        "required": ["amount", "callbackUrl", "description"]
      },
      "VerifyZarinpalPaymentDto": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string",
            "description": "Transaction authority",
            "example": "A00000000000000000000000000000000001"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount in IRR",
            "example": 100000
          }
        },
        "required": ["authority", "amount"]
      },
      "ProcessRefundZarinpalDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Transaction authority",
            "example": "A00000000000000000000000000000000001"
          },
          "amount": {
            "type": "number",
            "description": "Refund amount in IRR",
            "example": 50000
          },
          "description": {
            "type": "string",
            "description": "Refund description",
            "example": "Refund for order #123"
          },
          "method": {
            "type": "string",
            "description": "Refund method",
            "example": "bank_account"
          },
          "reason": {
            "type": "string",
            "description": "Reason for refund",
            "example": "Customer requested refund"
          }
        },
        "required": ["sessionId", "amount", "description", "method", "reason"]
      },
...
    }
  }
}